refactor queueSimulation
