+++ QueueSimulation/QueueSimulation/event.cpp
@@ -1,0 +1,9 @@
+//
+//  event.cpp
+//  Simulator
+//
+//  Created by Adam Quintana on 4/8/19.
+//  Copyright © 2019 Adam Quintana. All rights reserved.
+//
+
+#include "event.hpp"
